seems to work but sometimes with odd philo nb they keep dying, and still need to find out for this double dying issue

+++ philo/includes/proto.h
@@ -6,7 +6,7 @@
 /*   By: tcoeffet <tcoeffet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/21 14:45:20 by tcoeffet          #+#    #+#             */
-/*   Updated: 2025/03/28 13:21:29 by tcoeffet         ###   ########.fr       */
+/*   Updated: 2025/03/28 14:13:38 by tcoeffet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 
 void			*ft_calloc(size_t nmemb, size_t size);
 int				ft_atoi(const char *nptr);
-unsigned long	get_time(t_data *data);
+size_t			get_time(t_data *data);
 
 // PHILO
